refactor `caps_word` into `capslock`
