+++ app/dts/behaviors.dtsi
@@ -15,7 +15,7 @@
 #include <behaviors/bluetooth.dtsi>
 #include <behaviors/ext_power.dtsi>
 #include <behaviors/outputs.dtsi>
-#include <behaviors/caps_word.dtsi>
+#include <behaviors/capslock.dtsi>
 #include <behaviors/key_repeat.dtsi>
 #include <behaviors/backlight.dtsi>
 #include <behaviors/macros.dtsi>
