--- app/dts/behaviors/caps_word.dtsi
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2021 The ZMK Contributors
- *
- * SPDX-License-Identifier: MIT
- */
-
-#include <dt-bindings/zmk/keys.h>
-
-/ {
-	behaviors {
-		/omit-if-no-ref/ caps_word: behavior_caps_word {
-			compatible = "zmk,behavior-caps-word";
-			label = "CAPS_WORD";
-			#binding-cells = <0>;
-			bindings = <&kp CAPSLOCK>;
-			break-list = <SPACE TAB ENTER ESCAPE>;
-		};
-	};
-};
-
